update something for terry
