+++ chart.js
@@ -10,8 +10,27 @@ for (let i = 0; i <= maxDataPoints; i += 20) {
     labels.push(i.toString());
 }
 
-export function dataChartEvent(X,Y,Z){
-    chartData = [X, Y, Z];
+export function addData(X,Y,Z){
+    const chartData = [x, y, z]
+    if (flag) {
+        if (!(chartData.length == 0)) {
+            // 如果已經有100筆資料，則刪除第一筆資料
+            // console.log(dataPoints);
+            if (dataPoints >= maxDataPoints) {
+                myChart.data.datasets.forEach(dataset => {
+                    dataset.data.shift(); // 刪除第一筆資料
+                });
+            } else {
+                dataPoints++;
+            }
+
+            // 新增新的數據
+            myChart.data.datasets.forEach((dataset, index) => {
+                dataset.data.push(chartData[index]);
+            });
+        }
+        myChart.update(); // 更新圖表
+    }
 }
 
 export function creatNewChart() {
@@ -80,38 +99,38 @@ export function creatNewChart() {
             }
         }
     });
-    setInterval(() => {
-        if (flag) {
-            if (!(chartData.length == 0)) {
-                // 如果已經有100筆資料，則刪除第一筆資料
-                // console.log(dataPoints);
-                if (dataPoints >= maxDataPoints) {
-                    myChart.data.datasets.forEach(dataset => {
-                        dataset.data.shift(); // 刪除第一筆資料
-                    });
-                } else {
-                    dataPoints++;
-                }
+    // setInterval(() => {
+    //     if (flag) {
+    //         if (!(chartData.length == 0)) {
+    //             // 如果已經有100筆資料，則刪除第一筆資料
+    //             // console.log(dataPoints);
+    //             if (dataPoints >= maxDataPoints) {
+    //                 myChart.data.datasets.forEach(dataset => {
+    //                     dataset.data.shift(); // 刪除第一筆資料
+    //                 });
+    //             } else {
+    //                 dataPoints++;
+    //             }
 
-                // 新增新的數據
-                myChart.data.datasets.forEach((dataset, index) => {
-                    dataset.data.push(chartData[index]);
-                });
-            }
-            myChart.update(); // 更新圖表
-        }
-    }, 10);
+    //             // 新增新的數據
+    //             myChart.data.datasets.forEach((dataset, index) => {
+    //                 dataset.data.push(chartData[index]);
+    //             });
+    //         }
+    //         myChart.update(); // 更新圖表
+    //     }
+    // }, 10);
 }
 
 
 export function chartTypeEvent(){
     return chartType;
 }
-export function startChart(event){
+export function startChart(status){
     flag = event;
 }
 
-export function changeChart(event) {
+export function changeSensorChart(event) {
     chartType = event.target.value;
     myChart.data.datasets.forEach(dataset => {
         dataset.data = []
