+++ sw.js
@@ -3,15 +3,14 @@ self.addEventListener('install', (e) => {
     caches.open('fox-store').then((cache) => cache.addAll([
       '/white-cane-app/sytle.css',
       '/white-cane-app/index.js',
-      '/white-cane-app/animation_erase.js', 
-      '/white-cane-app/bluetooth.js', 
-      '/white-cane-app/chart.js', 
-      '/white-cane-app/csv_save.js', 
-      '/white-cane-app/decode.js', 
-      '/white-cane-app/keep_wake.js', 
-      '/white-cane-app/mouse_event.js', 
-      '/white-cane-app/querySelector_log.js', 
+      '/white-cane-app/animation_erase.js',
+      '/white-cane-app/bluetooth.js',
+      '/white-cane-app/chart.js',
+      '/white-cane-app/csv_save.js',
+      '/white-cane-app/keep_wake.js',
+      '/white-cane-app/mouse_event.js',
       '/white-cane-app/voice.js',
+      '/white-cane-app/utils.js',
     ])),
   );
 });
